Add db connection to Tasks and Groups

+++ Frontend/src/main/java/com/synapse/client/store/MembersStore.java
@@ -8,15 +8,15 @@ import java.util.Map;
 
 public class MembersStore {
     private static MembersStore instance;
-    private Map<Integer, ObservableList<User>> groupMembers;
+    private Map<Long, ObservableList<User>> groupMembers;
 
     private MembersStore() {
         groupMembers = new HashMap<>();
 
         ObservableList<User> usersGroup1 = FXCollections.observableArrayList();
-        usersGroup1.add(new User(1, "Ivan Ivanov", "ivan@test.com"));
-        usersGroup1.add(new User(2, "Petr Petrov", "petr@test.com"));
-        groupMembers.put(1, usersGroup1);
+        usersGroup1.add(new User(1L, "Ivan Ivanov", "ivan@test.com"));
+        usersGroup1.add(new User(2L, "Petr Petrov", "petr@test.com"));
+        groupMembers.put(1L, usersGroup1);
     }
 
     public static synchronized MembersStore getInstance() {
@@ -26,20 +26,19 @@ public class MembersStore {
         return instance;
     }
 
-    public ObservableList<User> getMembersByGroupId(int groupId) {
+    public ObservableList<User> getMembersByGroupId(Long groupId) {
         return groupMembers.computeIfAbsent(groupId, k -> FXCollections.observableArrayList());
     }
 
-    public void removeMember(int groupId, User user) {
+    public void removeMember(Long groupId, User user) {
         ObservableList<User> members = getMembersByGroupId(groupId);
         if (members != null) {
             members.remove(user);
-            // Тут будет запрос к серверу: server.kickUser(groupId, user.getId());
             System.out.println("User " + user.getUsername() + " kicked from group " + groupId);
         }
     }
 
-    public void addMember(int groupId, User user) {
+    public void addMember(Long groupId, User user) {
         getMembersByGroupId(groupId).add(user);
     }
 }
